Created and added mainLPTAnalysis.R file to folder
The purpose of this is to be used to include all of the analyses and post processing code for the LPT model output. This is the main shell that holds those 
modularized functions and calls upon the functions to analyze the data. 
It currently only includes the loading and unlisting for the modularized function ppAnalysis.R