Add ROMSTransects.R script
This is the first time the script ROMSTransects.R has been uploaded to this folder. 

It's purpose is to download, converts, and plot transects of oceanographic data from ROMS. The data are currently being downloaded and averaged daily. There is code to make plots of the individual daily averaged figures and a panel plot that contains all of the daily averages to visualize changes over time. 

This script uses functions from the OceanView package such as mapsigma that converts the transect matrix from sigma coordinates to depth values and increases the resolution.